Change uint 256 to uint.

+++ contracts/MyEpicNFT.sol
@@ -18,7 +18,7 @@ contract MyEpicNFT is ERC721URIStorage{
     using Counters for Counters.Counter;
     Counters.Counter private _tokenIds;
     //limit the number of NFTs to mint
-    uint256 private _max = 100;
+    uint private _max = 100;
 
     //SVG code. Keep this baseSvg variable and change the word to display 
     string baseSvg = "<svg xmlns='http://www.w3.org/2000/svg' preserveAspectRatio='xMinYMin meet' viewBox='0 0 350 350'><style>.base { fill: white; font-family: serif; font-size: 24px; }</style><rect width='100%' height='100%' fill='black' /><text x='50%' y='50%' class='base' dominant-baseline='middle' text-anchor='middle'>";
@@ -64,8 +64,8 @@ contract MyEpicNFT is ERC721URIStorage{
         return uint256(keccak256(abi.encodePacked(input)));
     }
 
-    function getNumberOfMintedNFTs() public view returns (uint256, uint256){
-        return (_tokenIds.current(), _max);
+    function getNumberOfMintedNFTs() public view returns (uint, uint){
+        return (uint(_tokenIds.current()), _max);
     }
 
     //a function that allows users to get an nft
